@@ -1,7 +1,7 @@
 {
   "name": "@org1/react-app",
   "scripts": {
-    "start": "webpack serve",
+    "start": "webpack serve  --mode=development",
     "start:standalone": "webpack serve --env standalone",
     "build": "webpack --mode=production",
     "analyze": "webpack --mode=production --env analyze",
@@ -18,48 +18,48 @@
     }
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/preset-react": "^7.12.7",
-    "@babel/runtime": "^7.12.5",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "babel-jest": "^26.6.3",
-    "concurrently": "^5.3.0",
+    "@babel/core": "^7.17.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@babel/runtime": "^7.17.8",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
+    "@types/testing-library__jest-dom": "^5.14.3",
+    "babel-jest": "^27.5.1",
+    "concurrently": "^7.0.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.15.0",
-    "eslint-config-prettier": "^7.0.0",
-    "eslint-plugin-prettier": "^3.2.0",
-    "husky": "^4.3.5",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.6.3",
-    "jest-cli": "^26.6.3",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0",
-    "single-spa-react": "^4.0.0",
-    "webpack": "^5.8.0",
-    "webpack-cli": "^4.2.0",
-    "webpack-config-single-spa-react": "^2.0.0",
-    "webpack-dev-server": "^4.0.0-beta.0",
-    "webpack-merge": "^5.4.0",
-    "@babel/preset-typescript": "^7.12.7",
-    "typescript": "^4.1.2",
-    "webpack-config-single-spa-ts": "^2.0.0",
-    "ts-config-single-spa": "^2.0.0",
-    "@types/testing-library__jest-dom": "^5.9.5",
+    "eslint": "^8.11.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-config-ts-react-important-stuff": "^3.0.0",
-    "webpack-config-single-spa-react-ts": "^2.0.0"
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.4",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^27.5.1",
+    "jest-cli": "^27.5.1",
+    "prettier": "^2.6.0",
+    "pretty-quick": "^3.1.3",
+    "react-router-dom": "^6.2.2",
+    "single-spa-react": "^4.6.1",
+    "ts-config-single-spa": "^3.0.0",
+    "typescript": "^4.6.3",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-config-single-spa-react": "^4.0.3",
+    "webpack-config-single-spa-react-ts": "^4.0.3",
+    "webpack-config-single-spa-ts": "^4.1.2",
+    "webpack-dev-server": "^4.7.4",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@types/jest": "^26.0.16",
-    "@types/react": "^16.9.32",
-    "@types/react-dom": "^16.9.6",
-    "@types/systemjs": "^6.1.0",
-    "@types/webpack-env": "^1.16.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0"
+    "@types/jest": "^27.4.1",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
+    "@types/systemjs": "^6.1.1",
+    "@types/webpack-env": "^1.16.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
