@@ -14,36 +14,38 @@
     }
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/runtime": "^7.12.5",
-    "concurrently": "^5.3.0",
+    "@babel/core": "^7.17.8",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-typescript": "^7.16.7",
+    "@babel/runtime": "^7.17.8",
+    "concurrently": "^7.0.0",
+    "copy-webpack-plugin": "^10.2.4",
     "cross-env": "^7.0.3",
-    "eslint": "^7.15.0",
-    "eslint-config-prettier": "^7.0.0",
-    "eslint-plugin-prettier": "^3.2.0",
-    "husky": "^4.3.5",
-    "jest": "^26.6.3",
-    "jest-cli": "^26.6.3",
-    "prettier": "^2.2.1",
-    "pretty-quick": "^3.1.0",
-    "serve": "^11.3.2",
-    "webpack": "^5.8.0",
-    "webpack-cli": "^4.2.0",
-    "webpack-dev-server": "^4.0.0-beta.0",
-    "webpack-merge": "^5.5.0",
-    "@babel/preset-typescript": "^7.12.7",
+    "eslint": "^8.11.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-config-ts-important-stuff": "^1.1.0",
-    "typescript": "^4.1.2",
-    "webpack-config-single-spa-ts": "^2.0.0",
-    "ts-config-single-spa": "^2.0.0"
+    "eslint-plugin-prettier": "^4.0.0",
+    "file-loader": "^6.2.0",
+    "husky": "^7.0.4",
+    "jest": "^27.5.1",
+    "jest-cli": "^27.5.1",
+    "prettier": "^2.6.0",
+    "pretty-quick": "^3.1.3",
+    "serve": "^13.0.2",
+    "ts-config-single-spa": "^3.0.0",
+    "typescript": "^4.6.3",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-config-single-spa-ts": "^4.1.2",
+    "webpack-dev-server": "^4.7.4",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@types/jest": "^26.0.16",
-    "@types/systemjs": "^6.1.0",
-    "single-spa": "^5.8.2",
-    "@types/webpack-env": "^1.16.0"
+    "@types/jest": "^27.4.1",
+    "@types/systemjs": "^6.1.1",
+    "@types/webpack-env": "^1.16.3",
+    "single-spa": "^5.9.3"
   }
 }
