@@ -3,8 +3,9 @@ import App from "./App";
 
 export default function Root(props) {
   return (
-    <section style={{ marginTop: 100 }}>
+    <section style={{ marginTop: 100, backgroundColor: "yellow" }}>
       <App name="Testapp" />
     </section>
   );
 }
+
