@@ -1,28 +1,38 @@
-const { merge } = require("webpack-merge");
-const singleSpaDefaults = require("webpack-config-single-spa-ts");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
+const { merge } = require('webpack-merge')
+const singleSpaDefaults = require('webpack-config-single-spa-ts')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
+const CopyPlugin = require('copy-webpack-plugin')
 
 module.exports = (webpackConfigEnv, argv) => {
-  const orgName = "org1";
+  
+  const orgName = 'org1'
   const defaultConfig = singleSpaDefaults({
     orgName,
-    projectName: "root-config",
+    projectName: 'root-config',
     webpackConfigEnv,
     argv,
     disableHtmlGeneration: true,
-  });
+  })
 
   return merge(defaultConfig, {
     // modify the webpack config however you'd like to by adding to this object
+    devServer: {
+      allowedHosts: [
+        'appdev.myebanking.net'
+      ],
+    },
     plugins: [
       new HtmlWebpackPlugin({
         inject: false,
-        template: "src/index.ejs",
+        template: 'src/index.ejs',
         templateParameters: {
           isLocal: webpackConfigEnv && webpackConfigEnv.isLocal,
           orgName,
         },
       }),
+      new CopyPlugin({
+        patterns: [{ from: 'public', to: '' }],
+      }),
     ],
-  });
-};
+  })
+}
